Rename things, add DeltaAnimation to abstract transitions
