--- internals/LED_data/Components/__init__.py
@@ -1,9 +1,0 @@
-from ._Animation import Animation
-from ._RGB import RGB
-from ._RGBArray import RGBArray
-
-__all__ = [
-    "Animation",
-    "RGB",
-    "RGBArray",
-]
