+++ internals/LED_data/_Animation/PulseAnimation.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from dataclasses import dataclass
 import math
 
-from ..Components import Animation, RGB, RGBArray
+from ..components import Animation, RGB, RGBArray
 
 @dataclass
 class PulseAnimation(Animation):
@@ -24,11 +24,6 @@ class PulseAnimation(Animation):
             strip[pixel] = current_color
 
         return strip
-
-    def pixel_state(self: PulseAnimation, pixel_id: int) -> RGB:
-        frame = self.frame()
-        fill_percentage = 0.5 * (1 - math.cos(2 * math.pi * frame))
-        return self.dark_led.interpolate(self.color, fill_percentage)
 
     def __str__(self) -> str:
         out = ""
