+++ internals/LED_data/components/_rgb.py
