+++ internals/LED_data/components/_rgb_array.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from rpi_ws281x import PixelStrip
 
 # from ._Animation.Animation import Animation
-from ._RGB import RGB
+from ._rgb import RGB
 
 class RGBArray():
     def __init__(self: RGBArray, array_length: int):
@@ -14,10 +14,6 @@ class RGBArray():
 
         for _ in range(array_length): 
             self._array.append(RGB(0,0,0))
-
-    # def update_strip_using(self: RGBArray, animation: Animation):
-    #     for pixel in range(self.size):
-    #         self._array[pixel] = animation.pixel_state(pixel)
 
     def send_output_to(self: RGBArray, strip: PixelStrip, update: bool = True):
         """Send the data from the RGB Array to the pixel_id strip output. \n
